test(igxChipsArea): Move helper function to UIInteraction functions. #6449
